testing infrastructure: separate the tests in several suites
Some tests are pretty long to run, namely the state machine and mantis
integration ones. On top of that, the mantis integration tests are
expecting the user to have all the Mantis binaries in $PATH, which
might not necessarily be the case.

We extract these tests into 3 tests suites:

- test: the "standard" one. It's meant to be run frequently. The tests
  are pretty fast to complete and do not require anything wild in $PATH.
- state-machine-tests: the state machine tests do not require anything
  wild in $PATH either, however, they are pretty long to complete (~4min).
- mantis-integration-tests: these integration tests require Mantis to
  be in $PATH on top of being pretty long to run (~2min).

We create a realease.nix job for each of these tests to make them run
on hydra.