Workaround for cabal trying to build dependencies in nix-shell
For source-repository-package's, cabal tries to build them on its own, even
when all dependencies are already provided by Nix. Relevant issues:
- https://github.com/haskell/cabal/issues/6049
- https://github.com/input-output-hk/ouroboros-network/pull/645
- https://github.com/haskell/cabal/issues/5586#issuecomment-479576084

This seems to be a problem even with a cabal that includes
https://github.com/haskell/cabal/pull/6917 (see
https://github.com/input-output-hk/haskell.nix/issues/720#issuecomment-745397468
for how to test a cabal-install 3.4)

The only known workaround is to remove the source-repository-package
sections from cabal.project, but this should only be done for cabal when
used from a nix-shell, not from cabal without a nix-shell, and not outside
the nix-shell.

To make this work smoothly, the script `scripts/nix-setup` can be used,
which splits the source-repository-package sections into cabal.project.srcs,
which is then again included from here (to make the Nix setup still work).
Running the script again undoes it.