Fix compilation of unit test suite
Notably this moves the inline encoding terms into binary files, without
changing them, and without adding any missing encoding tests.

This also adds types for the morpho node-to-node and node-to-client
versions in order for the test suite to be able to distinguish between
them. However node-to-client isn't necessary so will probably be removed
later, along with all other node-to-client bits