Fix unit tests, move morphoBlockSize out of morphoHeaderStd
morphoBlockSize is used to give an (over-)estimate of the block size.
But because it was inside morphoHeaderStd, this value had to be included
in the leader signature (morphoHeaderStd is the part that is signed).
This meant that there was no way to converge on a block size, since any
change to it introduced randomness to the encoding, which influences its
length since its a variable-size encoding. Moving it out of
morphoHeaderStd fixes this problem.

This only became a problem with the update because the new roundtrip_all
function checks that estimateBlockSize gives an overestimate.