Replace eTraceOpts with eTracers and bubble up cardano application
The Env type should be as general as possible. By passing eTracers
directly we can configure how each value should be traced, which could
be differently for tests.

As a consequence of this, the initialization of tracers was bubbled up
into the main application code and the conversion from configuration to
env

This removes almost all initialization done in the main morpho runner
functions