Nest iohk-monitoring config under Logging key
The iohk-monitoring configuration is now nested under the Logging key of
the config file. This includes settings like `minSeverity` and
`setupScribes`. The somewhat full list of options can be seen in
https://github.com/input-output-hk/iohk-monitoring-framework/blob/master/iohk-monitoring/test/config.yaml

The reason this is done is because mixing two different configuration
file parsers into the same namespace in the same file gets messy and confusing:
- There can be conflicts between fields. If iohk-monitoring suddenly introduces
  a field that morpho already uses, the value for morpho will be used
  for iohk-monitoring as well (which probably wouldn't make sense)
- The naming of options is inconsistent (iohk-monitoring uses lowercase,
  morpho uppercase)
- We can't provide error messages for unknown fields, as they might be
  used by the other config file parser

This is a breaking change, but the next commit will introduce an error
message that indicates what changes need to be done.