Apply bubblewrap patch for improved wrapped cabal repl experience
Running Ctrl-C in a cabal repl previously would exit the repl
completely, instead of just clearing the line. This is because Ctrl-C
sends a SIGINT to the bubblewrap process, which causes it to quit.

So this commit fixes this by applying a bubblewrap patch that forwards
SIGINT to the child process, allowing cabal repl to clear the line
properly on Ctrl-C.

In addition, this commit runs bubblewrap with exec to save a process