Remove unused socket file
The socket file is only used for node-to-client communication, of which
there is none for morpho. So having to specify a path for a socket has
absolutely no use.

In the previous commit, ouroboros-network was updated to include
https://github.com/input-output-hk/ouroboros-network/pull/2958, which
allows not specifying a socket at all. We're making use of this now.